+++ org/w3c/css/atrules/css/AtRuleImport.java
@@ -7,9 +7,9 @@
  * AtRuleMedia.java
  * $Id$
  */
-package org.w3c.css.parser;
+package org.w3c.css.atrules.css;
 
-import org.w3c.css.atrules.css.AtRuleMedia;
+import org.w3c.css.parser.AtRule;
 
 /**
  * This class manages all imports
