+++ org/w3c/css/parser/CssFouffa.java
@@ -10,6 +10,7 @@
 
 package org.w3c.css.parser;
 
+import org.w3c.css.atrules.css.AtRuleImport;
 import org.w3c.css.css.StyleSheetOrigin;
 import org.w3c.css.atrules.css.AtRuleMedia;
 import org.w3c.css.atrules.css.media.MediaFeature;
